kpe: Estimating the Value of Personalized Treatment Assignment

Point estimate, confidence interval, and one-sided p-value for the personalization effect – the expected outcome difference between the best personalized policy and the best single treatment applied to everyone. Implements the K-Fold Personalization Estimator of Li and Brunskill (2026). Supports binary and multi-arm discrete treatments, pluggable reward, contextual-policy, and best-arm learners, and is robust to mild misspecification of the plug-in models.

See also

Author

Maintainer: Emma Brunskill ebrun@cs.stanford.edu (ORCID)

Authors:

Emma Brunskill ebrun@cs.stanford.edu (ORCID)

Zhaoqi Li